pytorch problem stub
